Draft README
This provides at least a starter for a README for this lab. It doesn't say a tone, in part because there's a _lot_ in the code comments and in part because I don't have a lot of time. I'm not entirely sure what else I should document, but we'll see how the lab goes and make changes accordingly. My guess is that the basic algorithms will be one of the biggest sources of confusion, and we don't want to turn this README into a treatise on sorting algorithms.