Fix most of the clippy warnings
This fixes all but one of the warnings generated by `cargo clippy`. The one remaining warning is about an unused parameter, which they'll take care of when they implement things.

I also reduced the value of `size` and changed things so that they can actually `cargo run` the code and see the output in case that's helpful.